Attempting to roll net_connection into netcore to not have to keep track of 2 state machines. Might not actually work yet, needs testing.
